submit

+++ experiment/experiment/DEPSO.py
@@ -11,7 +11,6 @@ def sigmoid(x):
 
 
 def hybrid_DE_PSO_adjustment(GS, label_GS, TR, label_TR, Gm):
-    # sort the solutions and labels
     label_GS_sorted = np.sort(label_GS)
     GS_sorted = GS[np.argsort(label_GS)]
 
@@ -24,8 +23,8 @@ def hybrid_DE_PSO_adjustment(GS, label_GS, TR, label_TR, Gm):
     X_n, D = X.shape
     V = np.zeros((X_n, D))
 
-    X = X.astype(np.float64)  # 将 X 转换为 float64 类型
-    V = V.astype(np.float64)  # 确保 V 也是 float64 类型
+    X = X.astype(np.float64)
+    V = V.astype(np.float64)
 
     max_col_X = np.max(X, axis=0)
     min_col_X = np.min(X, axis=0)
@@ -33,14 +32,12 @@ def hybrid_DE_PSO_adjustment(GS, label_GS, TR, label_TR, Gm):
     class_n2 = len(np.unique(label_TR_sorted))
     class_n = max(class_n1, class_n2)
 
-    # create an instance of the KNN classifier
     KNN_model = KNeighborsClassifier(n_neighbors=3).fit(TR_sorted, label_TR_sorted)
     Predict_label = KNN_model.predict(X)
     Accuracy_global = np.sum(label_X == Predict_label) / X.shape[0]
 
     print('-----------------Start accuracy (should be 0):', Accuracy_global, '------------------------')
 
-    # 更新 correctIdx，不需要转换为列表
     correctIdx = np.zeros(len(label_GS_sorted))
     pos = np.where(label_X == Predict_label)[0]
     correctIdx[pos] = 1
@@ -102,7 +99,6 @@ def hybrid_DE_PSO_adjustment(GS, label_GS, TR, label_TR, Gm):
                 Fi = Fl + Fu * rand_Fi[0]
 
             KK = np.random.rand(1)
-            # PSO update: Calculate the new velocity
             if (i < len(label_TR)):
                 V[i] = np.multiply(sigmoid(V[i]), V[i]) \
                        + np.multiply(Fi * np.subtract(label_GS[i], X[i]), np.random.uniform()) \
@@ -110,7 +106,6 @@ def hybrid_DE_PSO_adjustment(GS, label_GS, TR, label_TR, Gm):
 
             X[i] += V[i]
 
-            # DE mutation
             if len(rand_r) >= 3:
                 U = X[i, :] + KK * (TR[rand_r[0] % TR.shape[0], :] - X[i, :] + Fi * (
                         TR[rand_r[2] % TR.shape[0], :] - TR[rand_r[3] % TR.shape[0], :]))
@@ -126,27 +121,21 @@ def hybrid_DE_PSO_adjustment(GS, label_GS, TR, label_TR, Gm):
             Position_less0 = np.where(U < min_col_X)
             U[Position_less0] = min_col_X[Position_less0]
 
-        # 找到需要更新的样本的索引
         incorrect_indices = np.where(correctIdx == 0)[0]
 
         for i in incorrect_indices:
-            # 如果 U 是一维数组，则直接赋值
             if U.ndim == 1:
                 X[i] = U
                 GS[i] = U
             else:
-                # 如果 U 是二维数组，则进行二维赋值
                 X[i, :] = U[i, :]
                 GS[i, :] = U[i, :]
 
-        # create KNN classification model
         KNN_model = KNeighborsClassifier(n_neighbors=3)
         KNN_model.fit(TR, label_TR)
-        # make predictions on new data
         Predict_label = KNN_model.predict(X)
         pos = np.where(label_X == Predict_label)
         Accuracy_U = sum(label_X == Predict_label) / X.shape[0];
-        # assume that the true labels are stored in `label_X` and the predicted labels are stored in `Predict_label`
         f1 = f1_score(label_X, Predict_label, average='macro')
         correctIdx[pos] = 1;
         if (Accuracy_U > Accuracy_global):
